requirements.txt
asgiref==3.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
decorator==5.1.1
Django==4.0.4
django-rest-knox==4.2.0
djangorestframework==3.13.1
dynaconf==3.1.8
hvac==0.11.2
idna==3.3
knox==0.1.14
loguru==0.6.0
Pillow==9.1.1
pycparser==2.21
pytz==2022.1
requests==2.27.1
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
validators==0.19.0
win32-setctime==1.1.0
